Add a limit of levels, meaning the game now has a proper ending
Due to the quirks in the way it was done, I had to create a dummy level
to be loaded before "winning". It add a couple of bytes, but nothing too
agravating.

The way it was done, it's not set in stone, but rather adding tools to
allow any implementation to do it on it's own best way.